create hook for initial sample load
